+++ ui/app/src/@types/mui.d.ts
@@ -26,6 +26,7 @@ declare module '@mui/material/styles' {
   interface Palette {
     secondaryContainer: Palette['primary']
     onSecondaryContainer: Palette['primary']
+    surfaceVariant: Palette['primary']
     onSurfaceVariant: Palette['primary']
     outline: Palette['primary']
   }
@@ -33,6 +34,7 @@ declare module '@mui/material/styles' {
   interface PaletteOptions {
     secondaryContainer: PaletteOptions['primary']
     onSecondaryContainer: PaletteOptions['primary']
+    surfaceVariant: PaletteOptions['primary']
     onSurfaceVariant: PaletteOptions['primary']
     outline: PaletteOptions['primary']
   }