+++ ui/app/src/components/FormField/AutocompleteField.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Chaos Mesh Authors.
+ * Copyright 2021 Chaos Mesh Authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,7 +21,7 @@ import MuiExtendsAutocompleteField from '@ui/mui-extends/esm/AutocompleteField'
 import type { AutocompleteFieldProps as MuiExtendsAutocompleteFieldProps } from '@ui/mui-extends/esm/AutocompleteField'
 import { T } from 'components/T'
 
-interface AutocompleteFieldProps extends MuiExtendsAutocompleteFieldProps {
+export interface AutocompleteFieldProps extends MuiExtendsAutocompleteFieldProps {
   name: string
 }