+++ ui/app/src/components/FormField/index.ts
@@ -17,6 +17,7 @@
 
 export { default as TextField } from './TextField'
 export { default as SelectField } from './SelectField'
-export { default as LabelField } from './LabelField'
 export { default as AutocompleteField } from './AutocompleteField'
+export { default as LabelField } from './LabelField'
+export { default as TextTextField } from './TextTextField'
 export { default as Submit } from './Submit'